Refactored project structure according to Python best practise.
