--- publish.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-source activate TwitterSave
-
-python src/python/main/publisher.py --configuration conf/
