--- run.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-source activate TwitterSave
-cat conf/banner.txt
-python src/python/main/saver.py --configuration conf/
