+++ scripts/publish.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+source activate TwitterSave
+
+python twitter_saver/publisher.py --configuration conf/
